split out ordering-party ui tests
